requirements.txt
pandas
sentence-transformers
faiss-cpu
requests